requirements.txt
numpy
numpy-stl
opencv-python